Metrics REST API tests
